fix(web): CI lint errors — extract chart currency fn, flatten nested ternary, data keys
