+++ web/src/features/dashboard/components/charts.tsx
@@ -41,6 +41,7 @@ import type {
   SiteDistributionSlice,
   SiteTrendPoint,
 } from '../types'
+import { formatChartCurrency } from './charts-currency'
 
 /** Number of ordinal colours exposed by the theme (--chart-1..5). */
 const PALETTE_SIZE = 5
@@ -53,20 +54,6 @@ function chartColor(index: number): string {
 /** Safe, stable dataKey for the i-th dynamic series (keeps CSS-var names valid). */
 function seriesKey(index: number): string {
   return `s${index}`
-}
-
-/**
- * Adaptive currency formatting for chart axes/tooltips. One precision per
- * magnitude band — the 6-decimal branch is gone because it rendered the zero
- * tick as "$0.000000" next to "$1.000" on the same axis (mixed precision
- * reads as a rounding bug). 3 decimals stays legible for sub-cent per-call
- * costs ($0.002) and keeps every tick on an axis in the same format.
- */
-export function formatChartCurrency(value: number): string {
-  if (!Number.isFinite(value)) return EM_DASH
-  const magnitude = Math.abs(value)
-  if (magnitude >= 1000) return `$${value.toFixed(2)}`
-  return `$${value.toFixed(3)}`
 }
 
 /** Percent-of-total string for the donut tooltip share row. */
